Separated card and actions, translations, clean-up
Code separated between a Decisions card and some insert operations that
belong in the new insert menu.
Translations have been added and used in the components.
