+++ app/components/besluit-context-card.js
@@ -1,0 +1,1 @@
+export { default } from '@lblod/ember-rdfa-editor-besluit-plugin/components/besluit-context-card';
